fix :wrong reject request exception

+++ src/main/java/com/toolc/stmap/domain/product/service/ProcessingRegisterRequestProduct.java
@@ -1,10 +1,13 @@
 package com.toolc.stmap.domain.product.service;
 
 import com.toolc.stmap.domain.product.entity.product.Product;
+import com.toolc.stmap.domain.product.exception.AlreadyPermittedProductException;
 import com.toolc.stmap.domain.product.exception.NotExistProductException;
 import com.toolc.stmap.domain.product.repository.interfaces.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import static com.toolc.stmap.global.Contracts.requires;
 
 @Service
 @RequiredArgsConstructor
@@ -21,6 +24,11 @@ public class ProcessingRegisterRequestProduct {
 
   public void reject(Long productId) {
     Product product = productRepository.findById(productId).orElseThrow(NotExistProductException::new);
+
+    if(product.getIsRegister()){
+      throw new AlreadyPermittedProductException();
+    }
+
     productRepository.delete(product);
   }
 }
